backend/requirements.txt
# Backend — Dependencias Python
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
requests==2.31.0
ortools==9.10.4067
python-dotenv==1.0.1
